Copy update for conversation deletion error (#3634)
* Move the delete group methods into a separate extension like the other actions

* Rename .delete action to .deleteGroup

* Update group deletion error copy

* Present alert on the top most view controller since the action controller target view controller doesn't exist when error occurs
