+++ Wire-iOS/Sources/UserInterface/GroupDetails/ConversationActions/ConversationActionController+ClearContent.swift
